Constructor and append working.
Built constructor for the Head and List  Node objects and a way to add an entry to the tail of the list. Also commented basic notes on how to build the other functions.

Functions added so far:
An element creator to display the list to easier visualize.

At function to report what is at a given index.

Size function to report length of a given linked list.

Append function to add a node to the tail of the list.